"TODO" directive for tests
A test can call the "todo()" function to mark itself as a stub test.
The test is only marked as in progress and the test function will
not be aborted. If the test function returns success status, either
the test actually works (and should no longer be todo), or there is
something wrong with the test.