Fix anchoring and alternation bugs and add tests
The end-of-string anchor `$` didn't work with expression
alternations. The problem was that for a pattern such as

    /^x?(xy|yx)*$/

when matched against e.g "xyyx", the first part `x?` matches
the initial "x", but the following group expression `xy|yx`, does
not match "yyx". Because the quantifier is zero or more it is,
however, accepted. The test for whether the whole string was matched
or not happened too late, and the pattern was reported as not
matching. What needed to happen was to back-track from the initial
match of "x", and that requires `match_exp` to know whether the
end of pattern and end of string should conicide.

Further, the alternation failed to match /^cc|ccc$/ against "ccc",
because the first matching branch was always chosen. Now `match_alt`
evaluates every branch and chooses one that matches the most
characters.