Turn align_to into an inline function
The result of C11's alignof operator is guaranteed to be a positive
power of two (§ 6.2.8 4). The align_to rounds an offset up to the
closest multiple of its argument, and can conventiently be performed
without explicit branches.

The function is_power_of_2 can be used to validate that the argument
to align_to is a power of two.