Add rescache module
The resource cache is a collection of resources (data, images, etc.)
that have been read from disk into a common internal format, and can
be retrieved using a key (which is usually a file name). For each
supported resource format there is a constructor that can turn it into
the common internal format, and a single destructor which frees up the
internal format.