Configure module directory with MODULE_DIR
The make file variable MODULE_DIR can be used to specify where source
modules are located, if not in the project root directory. E.g. all
modules can be put under a ./src/ directory, in which case you can
specify MODULE_DIR=src in either config.mk or as an environmental
variable.