Add TEST_MOD module configuration
The TEST_MOD variable specifies modules that a module's tests depend
upon. For instance, the module itself doesn't depend on a test module's
functions or for it to be even buildable, but only the tests do. The
TEST_MOD variable contain module names just like MOD, but they are only
used when compiling tests.