Make it possible to run tests in a tool
The test.sh script can be called with the -x and -o options, which
specify a binary that runs a test and options to pass to the test.

E.g. calling "./test.sh -x valgrind -o '-n'" runs a test with
valgrind, and passes the -n flag to each test.