Provide possibility for tests to be interactive
Tests in the ok module can be run in interactive mode for semi automated
tests. Tests that optionally can run in non-interactive mode can check
for the flag using the is_test_interactive() function. If a test only
makes sense to run interactively it should start off by calling
interactive_test() which skips the test if its not run interactively.