Make rescache accept arbitrary keys
Constructors in a resource cache are supposed to be able to create a
resource based on a key alone, and therefore it has to contain all the
information necessary for creation. When there are composite resources
the key might consist of more than one filename. In cases like these,
rather than parsing substrings from a serialized format, it is more
convenient to have a custom key with two simple nul-terminated strings.

The resource cache can now use keys other than strings. Users need to
pass the alignment of keys to the resource cache initializer and the
size of the key to the lookup function.

A key must be a preferably self-contained immutable value. It should at
least not contain pointers to non-static data. Because keys are compared
using memcmp, care must be taken when a key contains structure padding
but e.g. zeroing out the key first.