Fix garbled output in certain non-forking tests
When running a test binary without forking, if there were many test
functions, or if if printed a lot out text, then sometimes the output
was very messy. This was caused by buffered-IO not being flushed when
file descriptors were changed when standard output was re-directed.

The file streams has to be flushed before dup(2)ing and redirecting the
underlying file descriptors.