Split opengl/cache.c into several files
The OpenGL resource cache contains a collection of static assets that
can be used for rendering and might have been transferred to the GPU.
It uses several rescaches, one for each type of resource, and there can
be multiple loaders for each.

Putting them all into one file got a bit messy so here they all are in
separate files.

Summary:
 - Call things load/unload instead of make/free in rescache
 - Move make_*_cache, load_*, unload_*, and gl_load/release_* functions
   into separate files based on resource type
 - Handle resource keys in these files, hidden behind loaders
 - Separate files for resource convenience functions and loaders, e.g.
   shader.c and load-shader.c, to reduce coupling somewhat
 - Add shader program loader and test