Remove expected signal from test struct
The test definition structure used to have a field for specifying a
signal that is going to be raised unhandled. Signals are not available
on all platforms and need to be handled in a platform specific way, but
not through the portable test struct.

If a signal raised and is not handled in a test process it will now fail
the test. If it is required for tests to notify that a signal is going
to be raised, a better option might be to mark this fact by calling a
function at the start of a test, which informs the parent process of
this.