Retain tstack resources by destructor
The tstack used to consider pushing the same pointer twice to be a
programming error, and caused a failed assertion if this was not the
case. It sometimes useful, however, to push a pointer to a memory block
to be freed on failure, and to separately push the same pointer with a
destructor of the block's contents (which doesn't free the block).

To allow for this, it's not just a pointer, but a pointer-destructor
pair that needs to be unique. This requires that the tstack_retain
functions also receives the destructor to identify the pair.