Add hash map implementation
The hmap library in the new adt module is an implementation of hopscotch
hashing, which is similar to linear probing, except the length of the
linear scan is bounded by moving the closest free spot closer to the
desired location. The value to look up is usually found very close to
hashed index of the key.

The hash map does not only support pointers, but takes care of
allocation of members of arbitrary size and alignment, which makes its
usage similar to that of `malloc()`. It always makes copies of the keys
and is self-contained (except for references in the members). Keys need
to be zeroed-out before insertions and lookups, since they are compared
with `memcmp()`.

The hash function is the traditional Jenkin's 32bit hash.