Add test to measure load factor of hash tables
The hash table guarantees that a chain of buckets is bounded by a small
constant. The load_factor test checks how full a hash table becomes
before a rehash is needed. The hash table is rehashed when a chain
becomes too spread out, e.g. once it's no longer possible to append
another member to a chain, because its index is too far away from the
start of the chain. This ensures that lookups are fast, but it's
possible that insertions have linear or worse time complexity.

To reduce the time complexity for insertions, the linear probing for a
vacant position could be limited to only scan up to a constant number of
locations, and the number of retries, when moving the free location
could also have a limit.

For tables a little smaller than 1000000, the load factor was 99.5%,
with guarantees for time complexity of lookups.