Remove `buf` from `gl_make_xcontext()`
In the public API of the opengl module, there was a possibility to
provide the location to initialize the xstate into. This was a
module internal data structure, so it didn't make sense to use. Now the
interface is such that it's always dynamically allocated.

There is now a swap buffers function that is public.

Implemented using the new hash map, it's now possible to associate
an OpenGL drawable to an X drawable more than once.