Rename `wbuf_free` -> `wbuf_term`
The word "free" in the function name referred to freeing the buffer and
clearing out the wbuf struct. However, the name itself gives the
impression that the pointer passed to the function is also freed, which
is not the case.

From now on, use a consistent naming scheme, where *init*iators initiate
simple objects, which have already been allocated somewhere, and
*term*inators clear out and release any resources held by the object.

When the object is complex and its allocation and lifetime cannot easily
be handled by users of the object then use the word pairs *make*/*free*,
*load*/*unload*, *aquire*/*release*, and so on. The first one usually
implies a dynamic allocation followed by initialization. A call to the
second might not free the resources of the object immediately, in case
it is reference counted, for instance.