Expose definition of `struct hmap` in header
The `hmap` structure mainly keeps a pointer to the hash table table and
is very small and can be stack allocated or included on other structures
without a problem. This also improves memory locality and reduces
unnecessary memory fragmentation.

Remove the `hmap_make` and `hmap_free` functions which allocated the
hash map on the heap, and rename `hmap_deinit` to `hmap_term` to comply
with the new naming conventions.