Rename things in `opengl` module
Change names of functions, types, and files in the `opengl` module to be
more consistent in terms of initializing and allocating functions, but
also name spaces in functions.

I'm not certain about the pattern `foo_something_*`, for initiators and
terminators, however. For most functions that act on a `something` that
works well, but you don't call the `init` function on the `something`,
because it doesn't exist yet, so calling it `foo_init_something` might
sound better and be more logical. Same with termination functions, as
every now and then more context is needed to properly dispose of a data
structure than the data structure itself.