Always rebuild archives
When module object files are archived, don't update only the changed
files, but replace the existing one. First, GNU Make does not allow
parallel builds to concurrently manipulates archives, instead, the
behavior is undefined. Second, if a file has been removed it will not
be removed from the archive automatically.