Rename `glapi/opengl.h` -> `glapi/api.h`
Create a primary list of supported OpenGL core and extension functions
in `glapi/api.inc` in the form of the `GL_API` macro. Each entry
reflects a lazily initialized field in `struct gl_api`, and an
initialization and resolve function. Internally within the module, the
private `api.h` header can be used, but externally the pre-expanded
generated header with the same name should be used instead, which
mainly doesn't include the `GL_API` macro.