Add gap buffer implementation
A gap buffer is a generalization of a dynamic buffer (e.g wbuf) where
it's possible to efficiently insert content at places other than the
end. A gap buffer has a capacity C, which is the size of the dynamically
allocated memory block it references, a size S, which is the number of
bytes written, and a gap G, which is a range of bytes within the memory
block. The main invariant is C = S + |G|, which is basically the same
one as for a dynamic array, except that the gap doesn't have to be at
the end of the array.

The buffer uses four pointers: lbegin, lend, rbegin, and rend. The
`lbegin` pointer points to the memory block and `rend` points one byte
past it.  Pointer `lend` points to the first available byte in the gap
and `rbegin` at the the first unavailable byte so that
G = [lend, rbegin). The written content is the union of [lbegin, lend)
and [rbegin, rend), although either or both ranges can be empty.

The gap can be moved to any offset in the buffer and represents the
insertion point. When the gap gets filled the whole buffer is
re-allocated but the gap offset is maintained for future insertions.