Turn `struct gbuf` ranges into arrays
Rather than the `lbegin` to `lend` and `rbegin` to `rend` named ranges,
use two arrays `begin` and `end` of void pointers to specify the ranges.
This makes it slightly easier to iterate through the buffer no matter
where the gap is. The intended way to iterate over a gap buffer with
uniformly sized elements is now

    for (i = 0; i < 2; i++) {
        for (p = buf.begin[i]; p != buf.end[i]; p++) {
	    /* common body */
	}
    }

rather than using the `gbuf_get()` function, which makes offset
calculations tedious when iterating.