Add `gbuf_isuniform()`
Use this function to check whether a gap buffer is safe to use as a
container of uniformly sized objects. This implies that allocations have
been made using the same size throughout.

First of all, the sizes of both ranges must be muliples of the element
size. Second of all, the alignments of the arrays must be greater or
equal to the element alignment. The initial part of the buffer from
begin[0] to end[0] is always properly aligned for any data type, which
is a guarantee of `malloc(3)`. The alignment of the second part of the
buffer depends on the capacity of the buffer, since end[1] coincides
with the end of the buffer block. To support treating the second range
as a properly aligned array the capacity of the buffer is now always a
multiple of `alignof(max_align_t)`.

If the element alignment is equal to the maximum alignment then the
element size is a multiple thereof and so is the size of the gap.
If the element alignmet is less than the maximum then it is a power of
along the lines of `alignof(max_align_t) >> x` for some x. Such an
object is properly aligned if placed at the end of the buffer block.