Add function to statically allocate from buffers
Add functions similar to `alloc()`, `align()`, and `write()` to both
`wbuf` and `gbuf` which act similarly except that they fail instead of
attempting to re-allocate the underlying buffer when there's not enough
available space left.

In other words, it's possible to initially reserve a limited chuck of
space in the buffer and not have it grow arbitrarily large. In case of
a `wbuf`, this means that it can be used as a linear allocator because
pointers into the buffer remain valid when there is no resizing.