Add `gbuf_read()` and `gbuf_erase()`
The reader function can be used to copy an arbitrary range of bytes
from the gap buffer with logical offsets, no matter where the gap is
located.

The erase function is slightly easier to use than the previous
deletion and retraction functions, since it removes based on offset
and not only relative to current insert location.