Add binary heap procedures
Add functions for operating on binary heaps represented by arrays. All
of them require that the user keeps track of the size of the heap, the
comparison function, and the size of the elements manually. While they
can be used directly as is, when used in e.g. priority queues it might
be more convenient to wrap a dynamic array in a data structure with a
simplified interface with fewer parameters to pass.

The elements in the heap are ordered based on a comparison function
which compares two elements and returns a negative, zero, or positive
integer with the familiar semantics: a negative result implies that the
first argument is smaller than the second, by some metric. This is used
to order the elements and makes it a min-heap. If a max-heap is desired
then it is necesary to define a negated comparision function to use
instead.

- `bheap_init()` heapifies an array in linear time so that the root is
  stored at offset 0, its children at 1 and 2, and so on
- `bheap_insert()` adds an element to a heap with the assumption that
  there is capacity for one more element in the array than passed in as
  a parameter
- `bheap_remove()` removes an element from a heap, typically the first
  element (the minimum element), but it is technically possible to
  remove any element although this likely requires a linear search to
  find the element to remove