Add intrusive list
Intrusive structures, such as `struct ilist`, are supposed to be
included in other structs. Generic functions are provided that operate
on the included members directly, such as manipulating pointers, and
something like the `container_of` macro should be used to go from a node
to the containing structure.

A `struct ilist` is an intrusive linked list node. Linked lists are
either linear, which have a beginning and an end, or are circular.