Add fixed size memory pool `fixpool`
A `fixpool` controls a buffer that can be used to allocate blocks of
memory that have the same size. The size of the blocks need to have an
alignment and size equal to or greater than `void *`.

Available blocks in the buffer are linked together in a free list, the
links of which are stored within the available blocks.

It does not otherwise manage resources, i.e. it doesn't allocate or free
and buffers itself.