Add dynamic memory pool
`mempool`, in contrast to `fixpool`, allocates more memory for more
blocks on demand. For every N allocations it will allocate a new buffer
and split it into blocks that can be allocated.

Because pointer comparisons are strictly only defined for pointers
within the same object or array, there is no attempt at detecting
whether a pointer belongs to the pool when it is free'd or keeping track
of which buffers have been emptied and could be released. Instead once a
buffer has been allocated it will stay allocated until the pool itself
is terminated. This matches allocation patterns where there are as many
deallocations as allocations on average and no allocation spikes.