Add hierarchical transformation nodes
Add transformation hierarchies contained in `xylo_tgraph` which contain
nodes (`xylo_tnode`), where a node represents a transformation relative
to its parent. The type of transformation needs to be defined by the
user of the code.

The `xylo_tgraph` is a set of (root) transformations and handles
allocation and memory re-use. Based on the slides for of a 2009 talk by
Tony Albrecht titled "Pitfalls of Object Oriented Programming", which
describes a data-oriented approach to transformations, the
implementation is striving for memory locality for local and global
transformations in a cargo-cult manner.

Both `xylo_tgraph` and `xylo_tnode` are independent of the rest of the
xylo module and could be used handle transformations in both 2D and 3D
scenes. It could potentially be extracted into a more general module at
a later point.