Remove geometry shader
The use of the geometry shader allowed for a very compact representation
of shapes in terms of line loops which were expanded into a triangle fan
and curve triangles. This work is not essential for the drawing
algorithm overall.

Less work can be done by pre-computing the triangle fan and curve
triangle positions and attributes, at the cost of some redundancy and
space.

This commit removes the geometry shader and changes the format of the
vertex buffer so that it contains an index for triangles and vertices
with quadratic curve coordinates, instead of vertices for a line loop.