Split `xylo/private.h` into separate headers
The module internal interfaces have grown a bit and it's clearer to
split the header per source file.

The `types.h` is also a single file included by many sources that might
cause several re-compilations when changed, but it is nonetheless short
and provides a good overview of the module.