Use a single framebuffer for both quincunx samples
Use a framebuffer twice as wide as the target and render the center and
corner samples side by side. Since these samples don't overlap (which is
further enforced by clip planes) both samples can be rendered using
instanced rendering. Each instance is a sample and a uniform array of
vectors is used to position the sample. In addition to avoiding
switching framebuffers once, there are half as many state changes of
e.g. the stencil operations.