Add compacting garbage collection
Implement Cheney's breadth-first copying garbage collection algorithm.
Lists are made as compact as possible with cdr-coding (i.e. make pairs
adjacent and overlapping rather than a separate link field). Shared
structures is what prevents cdr-coding since changing the structure
might require other lists to be expanded and that makes the memory
requirements of the to-space to become greater than from-space.

Shared structure is identified with a pre-pass that counts the number of
references. Two bits are allocated for each cell; one bit for the first
visit and another for subsequent visits. This information is used during
the copying phase to ensure that what was adjacent in from-space remains
adjacent in to-space.

Normally list cells are allocated from the end of the allocation space
(from high towards low addresses). During collection lists are copied to
the low addresses of to-space so that lists can be traversed from the
head to tail. A side effect is that the root list is located at the
start of the region, which might could turn out to be useful in other
cases.

A to-space is one configuration of a struct lxspace which is not
suitable for use with e.g. lx_cons(). It must be converted back into an
allocation space after a semi-space collection is complete.