Change ref_offset indexing
ref_offset used to return cell+tag offset of a list element in order to
not waste two bits for every tag cell. This requires that all list
segments are aligned, however, and requires a division operation.

Simplify ref_offset to be the offset of the data cell (minus one)
instead so that tag cells don't have to be aligned.

The heap already reserved bits for the tag cells in any case.