A basic support for binary search trees
Add weight-balanced trees as a new fundamental type. Weight balanced
trees are self-balancing binary search trees which maintains the size of
the (sub-)tree in each node and is used to identify when a tree is
getting too unbalanced. Additionally, the size gives constant time
access to the size of a tree and gives you the nth largest item in the
tree in logarithmic time.

Trees in LX mimic association lists (a-lists) but provide logarithmic
rather than linear search time. Like a-lists, they are collections of
non-empty lists called entry lists, where the car of each entry list
contains its key. The cdr of the entry contains data associated with the
key, if any. Trees can represent either maps, multi-maps or sets
depending on use. Keys are unique and like all binary search trees, the
entries are sorted by their key.

An empty tree, like the empty list, is represented by a nil-reference.
Non-empty tree nodes are stored as lists in the heap, except that
references to tree nodes have a different type tag than references to
lists do. An internal tree node is a quadruple consisting of two
references to subtrees, the size of the tree, and a link to the entry
list. Equivalently, the subtree references and the size field form a
list of length three which is prepended to the entry list, i.e.

  (left-tree right-tree size . entry)

In a balanced binary tree, roughly half of the nodes are leaf nodes.
Leaf nodes can be identified from their size, which is 1, and always
have empty subtrees. Leaf nodes do not need to store these references,
which culls some overhead, but makes the size of leaf nodes different
from internal nodes.

The number of special cases are minimized for common operations, such
as lx_tree_size() and lx_tree_entry(), by having references to tree
nodes refer to the `size` field. Getting the size of tree node `t` is
simply `car(t)` and getting the entry is `cdr(t)`. If the size of a tree
node is one, then the lx_tree_left() and lx_tree_right() operations
return the empty tree, otherwise they retrieve the references just
before the size field. This scheme saves significant space but requires
some co-operation from the garbage collector.