Add lx_compare
Add generic comparator for comparing arbitrary LX values. Scalars are
sorted by their natural ordering, structures compare lexicographically.
When arguments are of different type, the type precedence is

    bool < number < string < list < tree

where integral floating point numbers compare equal. This is to mirror
the behavior of lx_equals where numbers of different types are compared
after being casted to floating point. Not sure if this is desirable
behavior or not.