Add set functions for trees based on join/split
Add lx_tree_union, lx_tree_isect, lx_tree_diff and lx_tree_filter based
on the *join* function. For trees of size *m* and *n* where m <= n
join-based tree algorithms run in O(m log (n/m + 1)).

These functions make the most sense when trees are sets, i.e. each entry
consists of keys only. The function union(a, b) adds all entries from b
to a, unless an entry with the same key already existed in `a`. For
set-like trees it is the union operator. For general trees it is a
shallow merge-left operator.

The function isect(a, b) and diff(a, b) only consider the keys of b and
respectively retains or removes entries of `a` with keys found in `b`.

The filter function keeps entries for which the predicate function
returns true.