Replace lx_modify() with lx_start() and lx_end()
Callback functions in C are not very ergonomic and hard to compose.
Instead make `struct lx_state` exposed in the API and flatten
`lx_modify()` and package the pre-modify and post-modify code into
`lx_start()` and `lx_end()`. These two delimit a block of code which
can be restarted and rather than a callback function, local variables
are in scope.

The same rules apply for code that within `lx_start()` and `lx_end()` as
as for the callback of `lx_modify()`.