Redesign the a-list trees as "p-list" maps
Association lists have the nice property that they can represent sets,
maps, and multi-maps, as well as non-homogenous maps. This also makes
them slightly tedious to work with since there's more structure to check
and validate before use.

Switch to plain maps instead where there's a single value associated
with each key. This is mostly a change to the API, where functions now
receive the key and value as separate parameters and generally returns
the value rather than the entry (as a list). The representation is the
same.

The read/written format looks like a p-list (alternating keys-values)
rather than a list of lists. The rules for writing structure textually
is a little more complex but is produces nicer results than before.